requirements.txt
appdirs==1.4.3
asgiref==3.2.3
certifi==2019.11.28
chardet==3.0.4
click==7.1.2
cs50==5.0.4
distlib==0.3.0
Django==3.0.3
filelock==3.0.12
Flask==1.1.2
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
Pillow==7.1.2
PyJWT==1.7.1
pytz==2019.3
requests==2.23.0
six==1.14.0
SQLAlchemy==1.3.17
sqlparse==0.3.0
termcolor==1.1.0
twilio==6.39.0
urllib3==1.25.9
virtualenv==20.0.4
Werkzeug==1.0.1
